+++ tests/e2e/main_test.go
@@ -45,7 +45,7 @@ func TestMain(m *testing.M) {
 	runID := envconf.RandomName("ns", 4)
 
 	kindClusterName := "mlp-e2e-tests"
-	kindImageName := "kindest/node:v1.32.2@sha256:f226345927d7e348497136874b6d207e0b32cc52154ad8323129352923a3142f"
+	kindImageName := "kindest/node:v1.32.5@sha256:e3b2327e3a5ab8c76f5ece68936e4cafaa82edf58486b769727ab0b3b97a5b0d"
 	if nameFromEnv, found := os.LookupEnv("KIND_NODE_IMAGE"); found {
 		kindImageName = nameFromEnv
 	}